Continue refactoring overlayManager -> neovimWindowManager
