+++ PLAN.md
@@ -1,1 +1,8 @@
-Rename OverlayManager - NeovimWindowManager & move to neovim
+'setWindowLineMapping'
+vs
+'setWindowDimensions'
+
+When is `currentWindowSizeChanged` called vs `windowDisplayUpdate`?
+What about `shouldMeasure`?
+
+Can these be tucked inside NeovimInstance?
