+++ browser/src/neovim/index.ts
@@ -3,4 +3,5 @@ export * from "./MsgPack"
 export * from "./Session"
 export * from "./NeovimProcessSpawner"
 export * from "./NeovimInstance"
+export * from "./NeovimWindowManager"
 export * from "./Window"
